requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.4.4
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-WTF==0.15.1
PyMySQL==1.0.2
python-dotenv==0.19.0
Werkzeug==2.0.1
email-validator==1.1.3
SQLAlchemy==1.3.24